# vl/api/openapi.yaml
openapi: 3.0.3
info:
  title: Serverless Nabaztag API
  description: An OpenAPI specification for controlling a Serverless Nabaztag smart rabbit.
  version: 1.1.0
  contact:
    url: https://github.com/andreax79/ServerlessNabaztag
servers:
  - url: http://{nabaztag_ip}
    description: Local Nabaztag device
    variables:
      nabaztag_ip:
        default: 192.168.1.100
        description: The IP address of your Nabaztag rabbit.
paths:
...
  /say:
    get:
      summary: Text to Speech
      tags: [Sound]
      description: Makes the Nabaztag speak the provided text.
      parameters:
        - name: t
          in: query
          required: true
          description: The text to be spoken.
          schema:
            type: string
      responses:
        '204':
          description: Command sent successfully.
...
  /ask:
    get:
      summary: Start a text AI turn
      tags: [AI]
      description: Sends text to the configured Realtime relay and plays the answer on the rabbit.
      parameters:
        - name: t
          in: query
          required: true
          description: Text to send to the voice assistant.
          schema:
            type: string
            maxLength: 4000
      responses:
        '204':
          description: AI turn started asynchronously.
...